+++ tests/score_tests/integration_tests/preprocessing_static/test_preprocessing_static.py
